requirements.txt
fastapi
google-auth
google-cloud-aiplatform
langchain
numpy==2.3.1
openai
pandas
pandas==2.3.1
pgvector
psycopg2-binary
pydantic
python-dateutil==2.9.0.post0
python-dotenv
pytz==2025.2
six==1.17.0
sqlalchemy
sqlalchemy>=2.0
tzdata==2025.2
uvicorn
